+++ crates/fluvio-hub-protocol/Cargo.toml
@@ -10,7 +10,6 @@ publish = false
 
 [dependencies]
 cargo_toml = { workspace = true }
-chrono = { workspace = true }
 const_format = { workspace = true }
 dirs = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
